class 및 struct 헤더파일로 분리
